+++ demisto_sdk/commands/validate/validate_manager.py
@@ -75,7 +75,7 @@ class ValidateManager:
             print_ignored_files=False, skip_conf_json=True, validate_id_set=False, file_path=None,
             validate_all=False, is_external_repo=False, skip_pack_rn_validation=False, print_ignored_errors=False,
             silence_init_prints=False, no_docker_checks=False, skip_dependencies=False, id_set_path=None, staged=False,
-            create_id_set=False, json_file_path=None, skip_schema_check=False, debug_git=False, ignore_untracked=False
+            create_id_set=False, json_file_path=None, skip_schema_check=False, debug_git=False, include_untracked=False
     ):
         # General configuration
         self.skip_docker_checks = False
@@ -94,7 +94,7 @@ class ValidateManager:
         self.staged = staged
         self.skip_schema_check = skip_schema_check
         self.debug_git = debug_git
-        self.ignore_untracked = ignore_untracked
+        self.include_untracked = include_untracked
 
         if json_file_path:
             self.json_file_path = os.path.join(json_file_path, 'validate_outputs.json') if \
@@ -968,13 +968,13 @@ class ValidateManager:
         # get files from git by status identification against prev-ver
         modified_files = self.git_util.modified_files(prev_ver=self.prev_ver,
                                                       committed_only=self.is_circle, staged_only=self.staged,
-                                                      debug=self.debug_git, ignore_untracked=self.ignore_untracked)
+                                                      debug=self.debug_git, include_untracked=self.include_untracked)
         added_files = self.git_util.added_files(prev_ver=self.prev_ver, committed_only=self.is_circle,
                                                 staged_only=self.staged, debug=self.debug_git,
-                                                ignore_untracked=self.ignore_untracked)
+                                                include_untracked=self.include_untracked)
         renamed_files = self.git_util.renamed_files(prev_ver=self.prev_ver, committed_only=self.is_circle,
                                                     staged_only=self.staged, debug=self.debug_git,
-                                                    ignore_untracked=self.ignore_untracked)
+                                                    include_untracked=self.include_untracked)
 
         # filter files only to relevant files
         filtered_modified, old_format_files = self.filter_to_relevant_files(modified_files)